refactor: remove unused schema, code
