+++ package.json
@@ -25,7 +25,7 @@
   },
   "dependencies": {
     "@giscus/react": "^2.2.6",
-    "@prisma/client": "^4.10.1",
+    "@prisma/client": "^4.11.0",
     "@radix-ui/react-toggle": "^1.0.1",
     "clsx": "^1.2.1",
     "contentlayer": "^0.3.0",
@@ -73,7 +73,7 @@
     "lint-staged": "^13.1.1",
     "postcss": "^8.4.21",
     "prettier": "latest",
-    "prisma": "^4.10.1",
+    "prisma": "^4.11.0",
     "seonest-post-gen": "^0.1.2",
     "tailwindcss": "^3.2.6",
     "tailwindcss-radix": "^2.7.0",
