+++ prisma/schema.prisma
@@ -11,44 +11,10 @@ datasource db {
   relationMode = "prisma"
 }
 
-model User {
-  id        Int        @id @default(autoincrement())
-  ipHash    String     @unique
-  code      String?
-  createdAt DateTime   @default(now())
-  updatedAt DateTime   @updatedAt
-  PostView  PostView[]
-  PostLike  PostLike[]
-}
-
 model Post {
   id        Int      @id @default(autoincrement())
   slug      String   @unique
   viewCount Int      @default(1)
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
-}
-
-model PostView {
-  id        Int      @id @default(autoincrement())
-  slug      String
-  viewer    User     @relation(fields: [viewerId], references: [id])
-  viewerId  Int
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-
-  @@index([slug])
-  @@index([viewerId])
-}
-
-model PostLike {
-  id        Int      @id @default(autoincrement())
-  slug      String
-  liker     User     @relation(fields: [likerId], references: [id])
-  likerId   Int
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-
-  @@index([slug])
-  @@index([likerId])
 }
