+++ src/components/PostViews/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './PostViews';
