--- src/services/user.service/index.ts
@@ -1,1 +1,0 @@
-export * from './user.service';
