--- src/services/user.service/types.ts
@@ -1,8 +1,0 @@
-interface UserParams {
-  ipHash: string;
-  code?: string;
-}
-
-export type CreateUserParams = UserParams;
-
-export type UpdateUserParams = UserParams;
